requirements.txt
PyPtt==1.0.9
python-dotenv==1.0.0
